Add CSV benchmarks and rowgroup benchmarks
Updated benchmarks to have CSV for both raw and opentap benchmarks
The benchmarks are now configured to be able to run for a much shorter
amount of time if the benchmarks are too long.

On top of this rowgroups have now been added. So a dataset can define
how many times it is supposed to be repeated.

Eventually i would like to move the repeated implementation into the
AllBenchmarks class, but for now it is defined within the
implementations themselves.

Also updated the CI/CD to be more specific instead of using the template
from OpenTAP/Parquet